backend/requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
black==24.4.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.8
daphne==4.1.2
dj-rest-auth==6.0.0
Django==5.0.4
django-environ==0.11.2
django-filter==24.2
django-redis==5.4.0
django-rest-swagger==2.2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
environ==1.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.3
Markdown==3.6
MarkupSafe==2.1.5
msgpack==1.0.8
mypy-extensions==1.0.0
norminette==3.3.53
openapi-codec==1.3.2
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
pip==24.0
platformdirs==4.2.1
pluggy==1.5.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
pytest==8.2.2
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
requests==2.31.0
service-identity==24.1.0
setuptools==69.5.1
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.1
uritemplate==4.1.1
urllib3==2.2.1
zope.interface==6.4.post2
